+++ Makefile
@@ -1,14 +1,11 @@
 NAME = minirt
 CC = gcc
 CFLAGS = -Wall -Wextra -Werror -MMD -MP
-CAMERA_SRCS = $(addprefix srcs/camera/, \
-		camera.c screen.c \
-		)
 IO_SRCS = $(addprefix srcs/io/, \
 		read_ambient.c read_camera.c read_cylinder.c read_file.c read_light.c read_plane.c read_sphere.c util.c validate_arg.c \
 		)
 OBJECTS_SRCS = $(addprefix srcs/objects/, \
-		object.c plane.c sphere.c cylinder.c cylinder_utils.c calc_lambert_cos_cylinder.c\
+		object.c plane.c sphere.c cylinder.c cylinder_utils.c calc_lambert_cos_cylinder.c rt.c camera.c screen.c\
 		)
 VECTOR3_SRCS = $(addprefix srcs/vector3/, \
 		add.c cos_of_angles.c inner_product.c multiply.c norm.c normalize.c outer_product.c print.c sub.c  vector3.c \
@@ -22,7 +19,7 @@ XLIBFT_SRCS = $(addprefix srcs/xlibft/, \
 UTILS_SRCS = $(addprefix srcs/utils/, \
 		error.c mlx_util.c util.c \
 		)
-SRCS = main.c $(CAMERA_SRCS) $(IO_SRCS) $(OBJECTS_SRCS) $(VECTOR3_SRCS) $(XSYSCALL_SRCS) $(XLIBFT_SRCS) $(UTILS_SRCS)
+SRCS = main.c $(IO_SRCS) $(OBJECTS_SRCS) $(VECTOR3_SRCS) $(XSYSCALL_SRCS) $(XLIBFT_SRCS) $(UTILS_SRCS)
 OBJS = $(SRCS:%.c=%.o)
 DEPENDS	= $(OBJS:.o=.d)
 LIBS = -lft -Llibft -lm
