+++ includes/minirt.h
@@ -6,7 +6,7 @@
 /*   By: takkatao <takkatao@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/02 16:29:44 by ahayashi          #+#    #+#             */
-/*   Updated: 2022/05/31 14:19:49 by takkatao         ###   ########.fr       */
+/*   Updated: 2022/05/31 14:54:46 by takkatao         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,5 @@
 # include "utils.h"
 # include "io.h"
 # include "objects.h"
-# include "camera.h"
 
 #endif
