+++ includes/objects.h
@@ -6,7 +6,7 @@
 /*   By: takkatao <takkatao@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/02 16:29:44 by ahayashi          #+#    #+#             */
-/*   Updated: 2022/05/31 13:59:36 by takkatao         ###   ########.fr       */
+/*   Updated: 2022/05/31 14:55:46 by takkatao         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,9 @@
 # include "xlibft.h"
 # include "define.h"
 # include "utils.h"
+
+t_vec3	*calc_camera_ray(t_scene *scene);
+t_vec3	*to_screen_vec(t_scene *scene, double x, double y);
 
 //sphere
 double	calc_t_sphere(t_ray *ray, t_sphere	*sphere);
@@ -32,10 +35,10 @@ double	calc_t_cylinder(t_ray *ray, t_cylinder	*cylinder);
 double	calc_height(t_vec3 *p_vec, t_cylinder	*cylinder);
 t_vec3	*create_p_vec(t_vec3 *s, t_vec3 *d, double t);
 
-
 //object
 double	calc_t(t_ray *ray, t_object *object);
 int		calc_ambient_light(t_ambient_light *a, t_object *object);
 int		calc_diffuse_light(t_ray *ray, t_object *object, t_light *light);
 
+int		calc_pixel_color(int x, int y, t_scene *sc);
 #endif
